+++ openvdb_cmd/vdb_tool/src/unittest.cpp
@@ -478,295 +478,295 @@ TEST_F(Test_vdb_tool, Stack)
     EXPECT_EQ(std::string("bla,foo,bar,bob"), ss.str());
 }// Stack
 
-TEST_F(Test_vdb_tool, Computer)
+TEST_F(Test_vdb_tool, Processor)
 {
     using namespace openvdb::vdb_tool;
-    Computer comp;
+    Processor proc;
 
     // test set and get, i.e. @ and $
-    EXPECT_THROW({comp("{$file}");}, std::invalid_argument);
-    EXPECT_THROW({comp("{dup}");},   std::invalid_argument);
-    EXPECT_THROW({comp("{drop}");},  std::invalid_argument);
-    EXPECT_THROW({comp("{swap}");},  std::invalid_argument);
+    EXPECT_THROW({proc("{$file}");}, std::invalid_argument);
+    EXPECT_THROW({proc("{dup}");},   std::invalid_argument);
+    EXPECT_THROW({proc("{drop}");},  std::invalid_argument);
+    EXPECT_THROW({proc("{swap}");},  std::invalid_argument);
 
-    EXPECT_EQ(0, comp.memory().size());
-    EXPECT_EQ(std::to_string(openvdb::math::pi<float>()), comp("{$pi}"));
-    EXPECT_EQ(std::to_string(openvdb::math::pi<float>()), comp("{pi:get}"));
-    EXPECT_EQ(std::to_string(2.718281828459), comp("{$e}"));
-    EXPECT_TRUE(comp("{path/base_0123.ext:@file}").empty());
-    EXPECT_EQ("path/base_0123.ext", comp("{$file}"));
-    EXPECT_TRUE(comp("{1:@G}").empty());
-    EXPECT_EQ("1", comp("{$G}"));
-    EXPECT_TRUE(comp("{$file:upper:@file2}").empty());
-    EXPECT_EQ("PATH/BASE_0123.EXT", comp("{$file2}"));
-    EXPECT_TRUE(comp("{$G:1000:+:@F}").empty());
-    EXPECT_EQ("1001", comp("{$F}"));
-    EXPECT_TRUE(comp("{0.1:@x:0.2:@y}").empty());
-    EXPECT_EQ("0.1", comp("{$x}"));
-    EXPECT_EQ("0.2", comp("{$y}"));
-    EXPECT_TRUE(comp("{1:$G:+:@G}").empty());
-    EXPECT_EQ("2", comp("{$G}"));
-    EXPECT_TRUE(comp("{$G:++:@G}").empty());
-    EXPECT_EQ("3", comp("{$G}"));
-    EXPECT_EQ(6, comp.memory().size());
+    EXPECT_EQ(0, proc.memory().size());
+    EXPECT_EQ(std::to_string(openvdb::math::pi<float>()), proc("{$pi}"));
+    EXPECT_EQ(std::to_string(openvdb::math::pi<float>()), proc("{pi:get}"));
+    EXPECT_EQ(std::to_string(2.718281828459), proc("{$e}"));
+    EXPECT_TRUE(proc("{path/base_0123.ext:@file}").empty());
+    EXPECT_EQ("path/base_0123.ext", proc("{$file}"));
+    EXPECT_TRUE(proc("{1:@G}").empty());
+    EXPECT_EQ("1", proc("{$G}"));
+    EXPECT_TRUE(proc("{$file:upper:@file2}").empty());
+    EXPECT_EQ("PATH/BASE_0123.EXT", proc("{$file2}"));
+    EXPECT_TRUE(proc("{$G:1000:+:@F}").empty());
+    EXPECT_EQ("1001", proc("{$F}"));
+    EXPECT_TRUE(proc("{0.1:@x:0.2:@y}").empty());
+    EXPECT_EQ("0.1", proc("{$x}"));
+    EXPECT_EQ("0.2", proc("{$y}"));
+    EXPECT_TRUE(proc("{1:$G:+:@G}").empty());
+    EXPECT_EQ("2", proc("{$G}"));
+    EXPECT_TRUE(proc("{$G:++:@G}").empty());
+    EXPECT_EQ("3", proc("{$G}"));
+    EXPECT_EQ(6, proc.memory().size());
 
     // test file-name methods
-    EXPECT_EQ("path", comp("{$file:path}"));
-    EXPECT_EQ("base_0123.ext", comp("{$file:file}"));
-    EXPECT_EQ("base_0123", comp("{$file:name}"));
-    EXPECT_EQ("base_", comp("{$file:base}"));
-    EXPECT_EQ("0123", comp("{$file:number}"));
-    EXPECT_EQ("ext", comp("{$file:ext}"));
+    EXPECT_EQ("path", proc("{$file:path}"));
+    EXPECT_EQ("base_0123.ext", proc("{$file:file}"));
+    EXPECT_EQ("base_0123", proc("{$file:name}"));
+    EXPECT_EQ("base_", proc("{$file:base}"));
+    EXPECT_EQ("0123", proc("{$file:number}"));
+    EXPECT_EQ("ext", proc("{$file:ext}"));
 
-    EXPECT_EQ("6", comp("{5:1:+}"));
-    EXPECT_EQ(std::to_string(6.0f), comp("{5.0:1:+}"));
-    EXPECT_EQ(std::to_string(6.2f), comp("{5.0:1.2:+}"));
+    EXPECT_EQ("6", proc("{5:1:+}"));
+    EXPECT_EQ(std::to_string(6.0f), proc("{5.0:1:+}"));
+    EXPECT_EQ(std::to_string(6.2f), proc("{5.0:1.2:+}"));
 
-    EXPECT_EQ("4", comp("{5:1:-}"));
-    EXPECT_EQ(std::to_string(4.0f), comp("{5.0:1:-}"));
-    EXPECT_EQ(std::to_string(3.8f), comp("{5.0:1.2:-}"));
+    EXPECT_EQ("4", proc("{5:1:-}"));
+    EXPECT_EQ(std::to_string(4.0f), proc("{5.0:1:-}"));
+    EXPECT_EQ(std::to_string(3.8f), proc("{5.0:1.2:-}"));
 
-    EXPECT_EQ("10", comp("{5:2:*}"));
-    EXPECT_EQ(std::to_string(10.0f), comp("{5.0:2:*}"));
-    EXPECT_EQ(std::to_string(6.0f), comp("{5.0:1.2:*}"));
+    EXPECT_EQ("10", proc("{5:2:*}"));
+    EXPECT_EQ(std::to_string(10.0f), proc("{5.0:2:*}"));
+    EXPECT_EQ(std::to_string(6.0f), proc("{5.0:1.2:*}"));
 
-    EXPECT_EQ("5", comp("{10:2:/}"));
-    EXPECT_EQ("0", comp("{2:10:/}"));
-    EXPECT_EQ(std::to_string(5.0f), comp("{10.0:2.0:/}"));
-    EXPECT_EQ(std::to_string(0.2f), comp("{2.0:10.0:/}"));
+    EXPECT_EQ("5", proc("{10:2:/}"));
+    EXPECT_EQ("0", proc("{2:10:/}"));
+    EXPECT_EQ(std::to_string(5.0f), proc("{10.0:2.0:/}"));
+    EXPECT_EQ(std::to_string(0.2f), proc("{2.0:10.0:/}"));
 
-    EXPECT_EQ("6", comp("{5:++}"));
-    EXPECT_EQ(std::to_string(6.2f), comp("{5.2:++}"));
+    EXPECT_EQ("6", proc("{5:++}"));
+    EXPECT_EQ(std::to_string(6.2f), proc("{5.2:++}"));
 
-    EXPECT_EQ("4", comp("{5:--}"));
-    EXPECT_EQ(std::to_string(4.2f), comp("{5.2:--}"));
+    EXPECT_EQ("4", proc("{5:--}"));
+    EXPECT_EQ(std::to_string(4.2f), proc("{5.2:--}"));
 
-    EXPECT_EQ("0", comp("{5:2:==}"));
-    EXPECT_EQ("0", comp("{5.0:2.0:==}"));
-    EXPECT_EQ("1", comp("{5:5:==}"));
-    EXPECT_EQ("1", comp("{5.0:5.0:==}"));
-    EXPECT_EQ("0", comp("{foo:bar:==}"));
-    EXPECT_EQ("1", comp("{foo:foo:==}"));
+    EXPECT_EQ("0", proc("{5:2:==}"));
+    EXPECT_EQ("0", proc("{5.0:2.0:==}"));
+    EXPECT_EQ("1", proc("{5:5:==}"));
+    EXPECT_EQ("1", proc("{5.0:5.0:==}"));
+    EXPECT_EQ("0", proc("{foo:bar:==}"));
+    EXPECT_EQ("1", proc("{foo:foo:==}"));
 
-    EXPECT_EQ("1", comp("{5:2:!=}"));
-    EXPECT_EQ("1", comp("{5.0:2.0:!=}"));
-    EXPECT_EQ("0", comp("{5:5:!=}"));
-    EXPECT_EQ("0", comp("{5.0:5.0:!=}"));
-    EXPECT_EQ("1", comp("{foo:bar:!=}"));
-    EXPECT_EQ("0", comp("{foo:foo:!=}"));
+    EXPECT_EQ("1", proc("{5:2:!=}"));
+    EXPECT_EQ("1", proc("{5.0:2.0:!=}"));
+    EXPECT_EQ("0", proc("{5:5:!=}"));
+    EXPECT_EQ("0", proc("{5.0:5.0:!=}"));
+    EXPECT_EQ("1", proc("{foo:bar:!=}"));
+    EXPECT_EQ("0", proc("{foo:foo:!=}"));
 
-    EXPECT_EQ("0", comp("{5:2:<=}"));
-    EXPECT_EQ("0", comp("{5.0:2.0:<=}"));
-    EXPECT_EQ("0", comp("{foo:bar:<=}"));
-    EXPECT_EQ("1", comp("{2:5:<=}"));
-    EXPECT_EQ("1", comp("{2.0:5.0:<=}"));
-    EXPECT_EQ("1", comp("{bar:foo:<=}"));
-    EXPECT_EQ("1", comp("{5:5:<=}"));
-    EXPECT_EQ("1", comp("{5.0:5.0:<=}"));
-    EXPECT_EQ("1", comp("{foo:foo:<=}"));
+    EXPECT_EQ("0", proc("{5:2:<=}"));
+    EXPECT_EQ("0", proc("{5.0:2.0:<=}"));
+    EXPECT_EQ("0", proc("{foo:bar:<=}"));
+    EXPECT_EQ("1", proc("{2:5:<=}"));
+    EXPECT_EQ("1", proc("{2.0:5.0:<=}"));
+    EXPECT_EQ("1", proc("{bar:foo:<=}"));
+    EXPECT_EQ("1", proc("{5:5:<=}"));
+    EXPECT_EQ("1", proc("{5.0:5.0:<=}"));
+    EXPECT_EQ("1", proc("{foo:foo:<=}"));
 
-    EXPECT_EQ("1", comp("{5:2:>=}"));
-    EXPECT_EQ("1", comp("{5.0:2.0:>=}"));
-    EXPECT_EQ("1", comp("{foo:bar:>=}"));
-    EXPECT_EQ("0", comp("{2:5:>=}"));
-    EXPECT_EQ("0", comp("{2.0:5.0:>=}"));
-    EXPECT_EQ("0", comp("{bar:foo:>=}"));
-    EXPECT_EQ("1", comp("{5:5:>=}"));
-    EXPECT_EQ("1", comp("{5.0:5.0:>=}"));
-    EXPECT_EQ("1", comp("{foo:foo:>=}"));
+    EXPECT_EQ("1", proc("{5:2:>=}"));
+    EXPECT_EQ("1", proc("{5.0:2.0:>=}"));
+    EXPECT_EQ("1", proc("{foo:bar:>=}"));
+    EXPECT_EQ("0", proc("{2:5:>=}"));
+    EXPECT_EQ("0", proc("{2.0:5.0:>=}"));
+    EXPECT_EQ("0", proc("{bar:foo:>=}"));
+    EXPECT_EQ("1", proc("{5:5:>=}"));
+    EXPECT_EQ("1", proc("{5.0:5.0:>=}"));
+    EXPECT_EQ("1", proc("{foo:foo:>=}"));
 
-    EXPECT_EQ("1", comp("{5:2:>}"));
-    EXPECT_EQ("1", comp("{5.0:2.0:>}"));
-    EXPECT_EQ("1", comp("{foo:bar:>}"));
-    EXPECT_EQ("0", comp("{2:5:>}"));
-    EXPECT_EQ("0", comp("{2.0:5.0:>}"));
-    EXPECT_EQ("0", comp("{bar:foo:>}"));
-    EXPECT_EQ("0", comp("{5:5:>}"));
-    EXPECT_EQ("0", comp("{5.0:5.0:>}"));
-    EXPECT_EQ("0", comp("{foo:foo:>}"));
+    EXPECT_EQ("1", proc("{5:2:>}"));
+    EXPECT_EQ("1", proc("{5.0:2.0:>}"));
+    EXPECT_EQ("1", proc("{foo:bar:>}"));
+    EXPECT_EQ("0", proc("{2:5:>}"));
+    EXPECT_EQ("0", proc("{2.0:5.0:>}"));
+    EXPECT_EQ("0", proc("{bar:foo:>}"));
+    EXPECT_EQ("0", proc("{5:5:>}"));
+    EXPECT_EQ("0", proc("{5.0:5.0:>}"));
+    EXPECT_EQ("0", proc("{foo:foo:>}"));
 
-    EXPECT_EQ("0", comp("{5:2:<}"));
-    EXPECT_EQ("0", comp("{5.0:2.0:<}"));
-    EXPECT_EQ("0", comp("{foo:bar:<}"));
-    EXPECT_EQ("1", comp("{2:5:<}"));
-    EXPECT_EQ("1", comp("{2.0:5.0:<}"));
-    EXPECT_EQ("1", comp("{bar:foo:<}"));
-    EXPECT_EQ("0", comp("{5:5:<}"));
-    EXPECT_EQ("0", comp("{5.0:5.0:<}"));
-    EXPECT_EQ("0", comp("{foo:foo:<}"));
+    EXPECT_EQ("0", proc("{5:2:<}"));
+    EXPECT_EQ("0", proc("{5.0:2.0:<}"));
+    EXPECT_EQ("0", proc("{foo:bar:<}"));
+    EXPECT_EQ("1", proc("{2:5:<}"));
+    EXPECT_EQ("1", proc("{2.0:5.0:<}"));
+    EXPECT_EQ("1", proc("{bar:foo:<}"));
+    EXPECT_EQ("0", proc("{5:5:<}"));
+    EXPECT_EQ("0", proc("{5.0:5.0:<}"));
+    EXPECT_EQ("0", proc("{foo:foo:<}"));
 
-    EXPECT_EQ("1", comp("{0:!}"));
-    EXPECT_EQ("0", comp("{1:!}"));
-    EXPECT_EQ("1", comp("{false:!}"));
-    EXPECT_EQ("0", comp("{true:!}"));
+    EXPECT_EQ("1", proc("{0:!}"));
+    EXPECT_EQ("0", proc("{1:!}"));
+    EXPECT_EQ("1", proc("{false:!}"));
+    EXPECT_EQ("0", proc("{true:!}"));
 
-    EXPECT_EQ("1", comp("{0:1:|}"));
-    EXPECT_EQ("1", comp("{1:0:|}"));
-    EXPECT_EQ("1", comp("{1:1:|}"));
-    EXPECT_EQ("0", comp("{0:0:|}"));
-    EXPECT_EQ("1", comp("{false:true:|}"));
-    EXPECT_EQ("0", comp("{false:false:|}"));
+    EXPECT_EQ("1", proc("{0:1:|}"));
+    EXPECT_EQ("1", proc("{1:0:|}"));
+    EXPECT_EQ("1", proc("{1:1:|}"));
+    EXPECT_EQ("0", proc("{0:0:|}"));
+    EXPECT_EQ("1", proc("{false:true:|}"));
+    EXPECT_EQ("0", proc("{false:false:|}"));
 
-    EXPECT_EQ("0", comp("{0:1:&}"));
-    EXPECT_EQ("0", comp("{1:0:&}"));
-    EXPECT_EQ("1", comp("{1:1:&}"));
-    EXPECT_EQ("0", comp("{0:0:&}"));
-    EXPECT_EQ("0", comp("{false:true:&}"));
-    EXPECT_EQ("0", comp("{false:false:&}"));
+    EXPECT_EQ("0", proc("{0:1:&}"));
+    EXPECT_EQ("0", proc("{1:0:&}"));
+    EXPECT_EQ("1", proc("{1:1:&}"));
+    EXPECT_EQ("0", proc("{0:0:&}"));
+    EXPECT_EQ("0", proc("{false:true:&}"));
+    EXPECT_EQ("0", proc("{false:false:&}"));
 
-    EXPECT_EQ("1", comp("{1:abs}"));
-    EXPECT_EQ("1", comp("{-1:abs}"));
-    EXPECT_EQ(std::to_string(1.2f), comp("{1.2:abs}"));
-    EXPECT_EQ(std::to_string(1.2f), comp("{-1.2:abs}"));
+    EXPECT_EQ("1", proc("{1:abs}"));
+    EXPECT_EQ("1", proc("{-1:abs}"));
+    EXPECT_EQ(std::to_string(1.2f), proc("{1.2:abs}"));
+    EXPECT_EQ(std::to_string(1.2f), proc("{-1.2:abs}"));
 
-    EXPECT_EQ(std::to_string(1.0f), comp("{1:ceil}"));
-    EXPECT_EQ(std::to_string(2.0f), comp("{1.2:ceil}"));
-    EXPECT_EQ(std::to_string(-1.0f), comp("{-1.2:ceil}"));
+    EXPECT_EQ(std::to_string(1.0f), proc("{1:ceil}"));
+    EXPECT_EQ(std::to_string(2.0f), proc("{1.2:ceil}"));
+    EXPECT_EQ(std::to_string(-1.0f), proc("{-1.2:ceil}"));
 
-    EXPECT_EQ(std::to_string(1.0f), comp("{1:floor}"));
-    EXPECT_EQ(std::to_string(1.0f), comp("{1.2:floor}"));
-    EXPECT_EQ(std::to_string(-2.0f), comp("{-1.2:floor}"));
+    EXPECT_EQ(std::to_string(1.0f), proc("{1:floor}"));
+    EXPECT_EQ(std::to_string(1.0f), proc("{1.2:floor}"));
+    EXPECT_EQ(std::to_string(-2.0f), proc("{-1.2:floor}"));
 
-    EXPECT_EQ("4", comp("{2:pow2}"));
-    EXPECT_EQ(std::to_string(4.0f), comp("{2.0:pow2}"));
+    EXPECT_EQ("4", proc("{2:pow2}"));
+    EXPECT_EQ(std::to_string(4.0f), proc("{2.0:pow2}"));
 
-    EXPECT_EQ("8", comp("{2:pow3}"));
-    EXPECT_EQ(std::to_string(8.0f), comp("{2.0:pow3}"));
+    EXPECT_EQ("8", proc("{2:pow3}"));
+    EXPECT_EQ(std::to_string(8.0f), proc("{2.0:pow3}"));
 
-    EXPECT_EQ("9", comp("{3:2:pow}"));
-    EXPECT_EQ(std::to_string(9.0f), comp("{3.0:2.0:pow}"));
+    EXPECT_EQ("9", proc("{3:2:pow}"));
+    EXPECT_EQ(std::to_string(9.0f), proc("{3.0:2.0:pow}"));
 
-    EXPECT_EQ("2", comp("{3:2:min}"));
-    EXPECT_EQ("-2", comp("{3:-2:min}"));
-    EXPECT_EQ(std::to_string(2.0f), comp("{3.0:2.0:min}"));
-    EXPECT_EQ(std::to_string(-2.0f), comp("{3.0:-2.0:min}"));
+    EXPECT_EQ("2", proc("{3:2:min}"));
+    EXPECT_EQ("-2", proc("{3:-2:min}"));
+    EXPECT_EQ(std::to_string(2.0f), proc("{3.0:2.0:min}"));
+    EXPECT_EQ(std::to_string(-2.0f), proc("{3.0:-2.0:min}"));
 
-    EXPECT_EQ("3", comp("{3:2:max}"));
-    EXPECT_EQ("3", comp("{3:-2:max}"));
-    EXPECT_EQ(std::to_string(3.0f), comp("{3.0:2.0:max}"));
-    EXPECT_EQ(std::to_string(2.0f), comp("{-3.0:2.0:max}"));
+    EXPECT_EQ("3", proc("{3:2:max}"));
+    EXPECT_EQ("3", proc("{3:-2:max}"));
+    EXPECT_EQ(std::to_string(3.0f), proc("{3.0:2.0:max}"));
+    EXPECT_EQ(std::to_string(2.0f), proc("{-3.0:2.0:max}"));
 
-    EXPECT_EQ("-3", comp("{3:neg}"));
-    EXPECT_EQ("3", comp("{-3:neg}"));
-    EXPECT_EQ(std::to_string(-3.0f), comp("{3.0:neg}"));
-    EXPECT_EQ(std::to_string(3.0f), comp("{-3.0:neg}"));
+    EXPECT_EQ("-3", proc("{3:neg}"));
+    EXPECT_EQ("3", proc("{-3:neg}"));
+    EXPECT_EQ(std::to_string(-3.0f), proc("{3.0:neg}"));
+    EXPECT_EQ(std::to_string(3.0f), proc("{-3.0:neg}"));
 
-    EXPECT_EQ(std::to_string(sin(2.0f)), comp("{2:sin}"));
-    EXPECT_EQ(std::to_string(sin(2.0f)), comp("{2.0:sin}"));
+    EXPECT_EQ(std::to_string(sin(2.0f)), proc("{2:sin}"));
+    EXPECT_EQ(std::to_string(sin(2.0f)), proc("{2.0:sin}"));
 
-    EXPECT_EQ(std::to_string(cos(2.0f)), comp("{2:cos}"));
-    EXPECT_EQ(std::to_string(cos(2.0f)), comp("{2.0:cos}"));
+    EXPECT_EQ(std::to_string(cos(2.0f)), proc("{2:cos}"));
+    EXPECT_EQ(std::to_string(cos(2.0f)), proc("{2.0:cos}"));
 
-    EXPECT_EQ(std::to_string(tan(2.0f)), comp("{2:tan}"));
-    EXPECT_EQ(std::to_string(tan(2.0f)), comp("{2.0:tan}"));
+    EXPECT_EQ(std::to_string(tan(2.0f)), proc("{2:tan}"));
+    EXPECT_EQ(std::to_string(tan(2.0f)), proc("{2.0:tan}"));
 
-    EXPECT_EQ(std::to_string(asin(2.0f)), comp("{2:asin}"));
-    EXPECT_EQ(std::to_string(asin(2.0f)), comp("{2.0:asin}"));
+    EXPECT_EQ(std::to_string(asin(2.0f)), proc("{2:asin}"));
+    EXPECT_EQ(std::to_string(asin(2.0f)), proc("{2.0:asin}"));
 
-    EXPECT_EQ(std::to_string(acos(2.0f)), comp("{2:acos}"));
-    EXPECT_EQ(std::to_string(acos(2.0f)), comp("{2.0:acos}"));
+    EXPECT_EQ(std::to_string(acos(2.0f)), proc("{2:acos}"));
+    EXPECT_EQ(std::to_string(acos(2.0f)), proc("{2.0:acos}"));
 
-    EXPECT_EQ(std::to_string(atan(2.0f)), comp("{2:atan}"));
-    EXPECT_EQ(std::to_string(atan(2.0f)), comp("{2.0:atan}"));
+    EXPECT_EQ(std::to_string(atan(2.0f)), proc("{2:atan}"));
+    EXPECT_EQ(std::to_string(atan(2.0f)), proc("{2.0:atan}"));
 
-    EXPECT_NEAR(openvdb::math::pi<float>(), str2float(comp("{180.0:d2r}")), 1e-4);
-    EXPECT_NEAR(180.0f, str2float(comp("{$pi:r2d}")), 1e-4);
+    EXPECT_NEAR(openvdb::math::pi<float>(), str2float(proc("{180.0:d2r}")), 1e-4);
+    EXPECT_NEAR(180.0f, str2float(proc("{$pi:r2d}")), 1e-4);
 
-    EXPECT_EQ(std::to_string(1.0f/2.0f), comp("{2:inv}"));
-    EXPECT_EQ(std::to_string(1.0f), comp("{1.0:inv}"));
-    EXPECT_EQ(std::to_string(1.0f/1.2f), comp("{1.2:inv}"));
+    EXPECT_EQ(std::to_string(1.0f/2.0f), proc("{2:inv}"));
+    EXPECT_EQ(std::to_string(1.0f), proc("{1.0:inv}"));
+    EXPECT_EQ(std::to_string(1.0f/1.2f), proc("{1.2:inv}"));
 
-    EXPECT_EQ(std::to_string(exp(1.2f)), comp("{1.2:exp}"));
-    EXPECT_EQ(std::to_string(log(1.2f)), comp("{1.2:ln}"));
-    EXPECT_EQ(std::to_string(log10(1.2f)), comp("{1.2:log}"));
-    EXPECT_EQ(std::to_string(sqrt(1.2f)), comp("{1.2:sqrt}"));
-    EXPECT_EQ("1", comp("{1:to_int}"));
-    EXPECT_EQ("1", comp("{1.2:to_int}"));
-    EXPECT_EQ(std::to_string(1.0f), comp("{1:to_float}"));
-    EXPECT_EQ(std::to_string(1.2f), comp("{1.2:to_float}"));
+    EXPECT_EQ(std::to_string(exp(1.2f)), proc("{1.2:exp}"));
+    EXPECT_EQ(std::to_string(log(1.2f)), proc("{1.2:ln}"));
+    EXPECT_EQ(std::to_string(log10(1.2f)), proc("{1.2:log}"));
+    EXPECT_EQ(std::to_string(sqrt(1.2f)), proc("{1.2:sqrt}"));
+    EXPECT_EQ("1", proc("{1:to_int}"));
+    EXPECT_EQ("1", proc("{1.2:to_int}"));
+    EXPECT_EQ(std::to_string(1.0f), proc("{1:to_float}"));
+    EXPECT_EQ(std::to_string(1.2f), proc("{1.2:to_float}"));
 
-    EXPECT_EQ("abcde012", comp("{AbCdE012:lower}"));
-    EXPECT_EQ("ABCDE012", comp("{AbCdE012:upper}"));
+    EXPECT_EQ("abcde012", proc("{AbCdE012:lower}"));
+    EXPECT_EQ("ABCDE012", proc("{AbCdE012:upper}"));
 
-    EXPECT_EQ("1", comp("{1:dup:==}"));
-    EXPECT_EQ("2", comp("{1:2:nip}"));
-    EXPECT_EQ("1", comp("{1:2:drop}"));
-    EXPECT_EQ(std::to_string(0.5f), comp("{1.0:2.0:/}"));
-    EXPECT_EQ(std::to_string(2.0f), comp("{1.0:2.0:swap:/}"));
-    EXPECT_EQ(std::to_string(2.0f/1.0f+1.0f), comp("{1.0:2.0:over:/:+}"));
+    EXPECT_EQ("1", proc("{1:dup:==}"));
+    EXPECT_EQ("2", proc("{1:2:nip}"));
+    EXPECT_EQ("1", proc("{1:2:drop}"));
+    EXPECT_EQ(std::to_string(0.5f), proc("{1.0:2.0:/}"));
+    EXPECT_EQ(std::to_string(2.0f), proc("{1.0:2.0:swap:/}"));
+    EXPECT_EQ(std::to_string(2.0f/1.0f+1.0f), proc("{1.0:2.0:over:/:+}"));
 
-    EXPECT_EQ(std::to_string(2.0f/3.0f+1.0f), comp("{1.0:2.0:3.0:/:+}"));
-    EXPECT_EQ(std::to_string(3.0f/1.0f+2.0f), comp("{1.0:2.0:3.0:rot:/:+}"));// rot(1 2 3) = 2 3 1
-    EXPECT_EQ(std::to_string(1.0f/2.0f+3.0f), comp("{1.0:2.0:3.0:tuck:/:+}"));// tuck(1 2 3) = 3 1 2
+    EXPECT_EQ(std::to_string(2.0f/3.0f+1.0f), proc("{1.0:2.0:3.0:/:+}"));
+    EXPECT_EQ(std::to_string(3.0f/1.0f+2.0f), proc("{1.0:2.0:3.0:rot:/:+}"));// rot(1 2 3) = 2 3 1
+    EXPECT_EQ(std::to_string(1.0f/2.0f+3.0f), proc("{1.0:2.0:3.0:tuck:/:+}"));// tuck(1 2 3) = 3 1 2
 
-    EXPECT_EQ("123", comp("{123:0:pad0}"));
-    EXPECT_EQ("123", comp("{123:1:pad0}"));
-    EXPECT_EQ("123", comp("{123:2:pad0}"));
-    EXPECT_EQ("123", comp("{123:3:pad0}"));
-    EXPECT_EQ("0123", comp("{123:4:pad0}"));
-    EXPECT_EQ("00123", comp("{123:5:pad0}"));
-    EXPECT_EQ("000123", comp("{123:6:pad0}"));
+    EXPECT_EQ("123", proc("{123:0:pad0}"));
+    EXPECT_EQ("123", proc("{123:1:pad0}"));
+    EXPECT_EQ("123", proc("{123:2:pad0}"));
+    EXPECT_EQ("123", proc("{123:3:pad0}"));
+    EXPECT_EQ("0123", proc("{123:4:pad0}"));
+    EXPECT_EQ("00123", proc("{123:5:pad0}"));
+    EXPECT_EQ("000123", proc("{123:6:pad0}"));
 
-    EXPECT_EQ("0", comp("{depth}"));
-    EXPECT_EQ("1", comp("{0:depth:scrape}"));
-    EXPECT_EQ("2", comp("{0:1:depth:scrape}"));
-    EXPECT_EQ("3", comp("{0:1:2:depth:scrape}"));
-    EXPECT_EQ("4", comp("{0:1:2:3:depth:scrape}"));
-    EXPECT_EQ("4", comp("{0:1:2:3:clear:4}"));
-    EXPECT_EQ("4", comp("{0:1:2:3:depth:@size:clear:$size}"));
+    EXPECT_EQ("0", proc("{depth}"));
+    EXPECT_EQ("1", proc("{0:depth:scrape}"));
+    EXPECT_EQ("2", proc("{0:1:depth:scrape}"));
+    EXPECT_EQ("3", proc("{0:1:2:depth:scrape}"));
+    EXPECT_EQ("4", proc("{0:1:2:3:depth:scrape}"));
+    EXPECT_EQ("4", proc("{0:1:2:3:clear:4}"));
+    EXPECT_EQ("4", proc("{0:1:2:3:depth:@size:clear:$size}"));
 
-    EXPECT_EQ("0", comp("{e:is_set}"));
-    EXPECT_EQ("0", comp("{pi:is_set}"));
-    EXPECT_EQ("0", comp("{foo:is_set}"));
-    EXPECT_EQ("1", comp("{8:@bar:bar:is_set}"));
+    EXPECT_EQ("0", proc("{e:is_set}"));
+    EXPECT_EQ("0", proc("{pi:is_set}"));
+    EXPECT_EQ("0", proc("{foo:is_set}"));
+    EXPECT_EQ("1", proc("{8:@bar:bar:is_set}"));
 
-    EXPECT_EQ(std::to_string(sqrt(0.1f*0.1f + 0.2f*0.2f)), comp("{$x:pow2:$y:pow2:+:sqrt}"));
+    EXPECT_EQ(std::to_string(sqrt(0.1f*0.1f + 0.2f*0.2f)), proc("{$x:pow2:$y:pow2:+:sqrt}"));
 
-    EXPECT_EQ("4",comp("{1:2:<:if(1:3:+)}"));
-    EXPECT_EQ("",comp("{1:2:>:if(1:3:+)}"));
-    EXPECT_EQ("1",comp("{5:@a:1:2:<:if(1:@a):$a}"));
-    EXPECT_EQ("5",comp("{5:@a:1:2:>:if(1:@a):$a}"));
+    EXPECT_EQ("4",proc("{1:2:<:if(1:3:+)}"));
+    EXPECT_EQ("",proc("{1:2:>:if(1:3:+)}"));
+    EXPECT_EQ("1",proc("{5:@a:1:2:<:if(1:@a):$a}"));
+    EXPECT_EQ("5",proc("{5:@a:1:2:>:if(1:@a):$a}"));
 
-    EXPECT_EQ("4",comp("{1:2:<:if(1:3:+?2:2:-)}"));
-    EXPECT_EQ("0",comp("{1:2:>:if(1:3:+?2:2:-)}"));
-    EXPECT_EQ("1",comp("{1:2:<:if(1:@a?2:@a):$a}"));
-    EXPECT_EQ("2",comp("{1:2:>:if(1:@a?2:@a):$a}"));
-    EXPECT_EQ(std::to_string(sqrt(4+16)),comp("{$pi:2:>:if(2:pow2:4:pow2:+:sqrt?2:sin)}"));
-    EXPECT_EQ(std::to_string(sin(2)),comp("{$pi:2:<:if(2:pow2:4:pow2:+:sqrt?2:sin)}"));
+    EXPECT_EQ("4",proc("{1:2:<:if(1:3:+?2:2:-)}"));
+    EXPECT_EQ("0",proc("{1:2:>:if(1:3:+?2:2:-)}"));
+    EXPECT_EQ("1",proc("{1:2:<:if(1:@a?2:@a):$a}"));
+    EXPECT_EQ("2",proc("{1:2:>:if(1:@a?2:@a):$a}"));
+    EXPECT_EQ(std::to_string(sqrt(4+16)),proc("{$pi:2:>:if(2:pow2:4:pow2:+:sqrt?2:sin)}"));
+    EXPECT_EQ(std::to_string(sin(2)),proc("{$pi:2:<:if(2:pow2:4:pow2:+:sqrt?2:sin)}"));
 
-    EXPECT_EQ("a", comp("{1:switch(1:a?2:b?3:c)}"));
-    EXPECT_EQ("b", comp("{2:switch(1:a?2:b?3:c)}"));
-    EXPECT_EQ("c", comp("{3:switch(1:a?2:b?3:c)}"));
-    //EXPECT_THROW({comp("{0:switch(1:a?2:b?3:c)}");}, std::invalid_argument);
-    //EXPECT_THROW({comp("{4:switch(1:a?2:b?3:c)}");}, std::invalid_argument);
-    EXPECT_EQ("SUPER", comp("{1:switch(1:super:upper?2:1:2:+?3:$pi)}"));
-    EXPECT_EQ("3", comp("{2:switch(1:super:upper?2:1:2:+?3:$pi)}"));
-    EXPECT_EQ(std::to_string(openvdb::math::pi<float>()), comp("{3:switch(1:super:upper?2:1:2:+?3:$pi)}"));
+    EXPECT_EQ("a", proc("{1:switch(1:a?2:b?3:c)}"));
+    EXPECT_EQ("b", proc("{2:switch(1:a?2:b?3:c)}"));
+    EXPECT_EQ("c", proc("{3:switch(1:a?2:b?3:c)}"));
+    //EXPECT_THROW({proc("{0:switch(1:a?2:b?3:c)}");}, std::invalid_argument);
+    //EXPECT_THROW({proc("{4:switch(1:a?2:b?3:c)}");}, std::invalid_argument);
+    EXPECT_EQ("SUPER", proc("{1:switch(1:super:upper?2:1:2:+?3:$pi)}"));
+    EXPECT_EQ("3", proc("{2:switch(1:super:upper?2:1:2:+?3:$pi)}"));
+    EXPECT_EQ(std::to_string(openvdb::math::pi<float>()), proc("{3:switch(1:super:upper?2:1:2:+?3:$pi)}"));
 
-    EXPECT_EQ("a", comp("{a:squash}"));
-    EXPECT_EQ("a,b,c,d", comp("{a:b:c:d:squash}"));
+    EXPECT_EQ("a", proc("{a:squash}"));
+    EXPECT_EQ("a,b,c,d", proc("{a:b:c:d:squash}"));
 
-    EXPECT_EQ("1", comp("{a:length}"));
-    EXPECT_EQ("3", comp("{foo:length}"));
-    EXPECT_EQ("7", comp("{foo bar:length}"));
+    EXPECT_EQ("1", proc("{a:length}"));
+    EXPECT_EQ("3", proc("{foo:length}"));
+    EXPECT_EQ("7", proc("{foo bar:length}"));
 
-    EXPECT_EQ("foobar", comp("{foo:bar:append}"));
+    EXPECT_EQ("foobar", proc("{foo:bar:append}"));
 
-    EXPECT_EQ("3", comp("{1,2,3:,:tokenize:depth:scrape}"));
-    EXPECT_EQ("5", comp("{1,2,3:,:tokenize:+:*}"));
+    EXPECT_EQ("3", proc("{1,2,3:,:tokenize:depth:scrape}"));
+    EXPECT_EQ("5", proc("{1,2,3:,:tokenize:+:*}"));
 
     // find two real roots of a quadratic polynomial
-    EXPECT_EQ("0.683375,7.316625", comp("{1:@a:-8:@b:5:@c:$b:pow2:4:$a:*:$c:*:-:@c:-2:$a:*:@a:$c:0:==:if($b:$a:/):$c:0:>:if($c:sqrt:dup:$b:+:$a:/:$b:rot:-:$a:/):squash}"));
+    EXPECT_EQ("0.683375,7.316625", proc("{1:@a:-8:@b:5:@c:$b:pow2:4:$a:*:$c:*:-:@c:-2:$a:*:@a:$c:0:==:if($b:$a:/):$c:0:>:if($c:sqrt:dup:$b:+:$a:/:$b:rot:-:$a:/):squash}"));
 
-    EXPECT_EQ("foo bar bla", comp("{foo_bar_bla:_: :replace}"));
-    EXPECT_EQ("foo_bar_bla", comp("{foo bar bla: :_:replace}"));
-    EXPECT_EQ("a b c d", comp("{a,b,c,d:,: :replace}"));
-    EXPECT_EQ("a b c d", comp("{a:b:c:d:squash:,: :replace}"));
+    EXPECT_EQ("foo bar bla", proc("{foo_bar_bla:_: :replace}"));
+    EXPECT_EQ("foo_bar_bla", proc("{foo bar bla: :_:replace}"));
+    EXPECT_EQ("a b c d", proc("{a,b,c,d:,: :replace}"));
+    EXPECT_EQ("a b c d", proc("{a:b:c:d:squash:,: :replace}"));
 
-    EXPECT_EQ("_bar", comp("{foo_bar:foo:erase}"));
-    EXPECT_EQ("f_bar", comp("{foo_bar:o:erase}"));
-    EXPECT_EQ("foobar", comp("{foo bar: :erase}"));
-}// Computer
+    EXPECT_EQ("_bar", proc("{foo_bar:foo:erase}"));
+    EXPECT_EQ("f_bar", proc("{foo_bar:o:erase}"));
+    EXPECT_EQ("foobar", proc("{foo bar: :erase}"));
+}// Processor
 
 TEST_F(Test_vdb_tool, ToolParser)
 {